+++ src/Subjects/AbstractSubject.php
@@ -268,7 +268,8 @@ abstract class AbstractSubject implements SubjectInterface, FilesystemSubjectInt
         EntityTypeCodes::CATALOG_PRODUCT_PRICE         => EntityTypeCodes::CATALOG_PRODUCT,
         EntityTypeCodes::CATALOG_PRODUCT_INVENTORY     => EntityTypeCodes::CATALOG_PRODUCT,
         EntityTypeCodes::CATALOG_PRODUCT_INVENTORY_MSI => EntityTypeCodes::CATALOG_PRODUCT,
-        EntityTypeCodes::CATALOG_PRODUCT_TIER_PRICE    => EntityTypeCodes::CATALOG_PRODUCT
+        EntityTypeCodes::CATALOG_PRODUCT_TIER_PRICE    => EntityTypeCodes::CATALOG_PRODUCT,
+        EntityTypeCodes::CUSTOMER                      => EntityTypeCodes::CUSTOMER
     );
 
     /**